requirements.txt
pandas==1.5.3
numpy==1.23.5
advent-of-code-data==1.3.2
pyyaml==6.0